+++ packages/manager/apps/vrack-services/src/pages/endpoints/Create.tsx
@@ -7,8 +7,8 @@ import {
   ODS_SELECT_SIZE,
   OdsSelectValueChangeEvent,
 } from '@ovhcloud/ods-components';
-import { useOvhTracking } from '@ovh-ux/manager-react-shell-client';
 import { ApiError, ApiResponse } from '@ovh-ux/manager-core-api';
+import { PageType, useOvhTracking } from '@ovh-ux/manager-react-shell-client';
 import {
   VrackServices,
   getVrackServicesResourceListQueryKey,
@@ -22,15 +22,14 @@ import {
   serviceTypeSelectName,
   serviceNameSelectName,
   getEndpointCreationMutationKey,
-} from './constants';
+} from './endpoints.constants';
 import { useServiceList, useVrackService } from '@/utils/vs-utils';
 import { ErrorPage } from '@/components/Error';
 import { FormField } from '@/components/FormField';
-import { urls, pageTrackingLabels } from '@/router/constants';
+import { urls } from '@/router/constants';
+import { PageName } from '@/utils/tracking';
 
-const dataTrackingPath = pageTrackingLabels[urls.createEndpoint];
-
-const EndpointCreationPage: React.FC = () => {
+export default function EndpointCreationPage() {
   const { t } = useTranslation('vrack-services/endpoints');
   const { id } = useParams();
   const [serviceType, setServiceType] = React.useState<string | undefined>(
@@ -88,12 +87,18 @@ const EndpointCreationPage: React.FC = () => {
         queryClient.invalidateQueries({
           queryKey: getVrackServicesResourceQueryKey(id),
         }),
-        trackPage({ path: dataTrackingPath, value: '-success' }),
+        trackPage({
+          pageName: PageName.pendingCreateEndpoint,
+          pageType: PageType.bannerInfo,
+        }),
       ]);
       navigate(dashboardUrl);
     },
     onError: () => {
-      trackPage({ path: dataTrackingPath, value: '-error' });
+      trackPage({
+        pageName: PageName.errorCreateEndpoint,
+        pageType: PageType.bannerError,
+      });
     },
   });
 
@@ -121,11 +126,10 @@ const EndpointCreationPage: React.FC = () => {
       title={t('createPageTitle')}
       description={t('createPageDescription')}
       createButtonLabel={t('createEndpointButtonLabel')}
-      dataTrackingPath={dataTrackingPath}
-      createButtonDataTracking="::confirm"
       formErrorMessage={t('endpointCreationError', {
         error: error?.response.data.message,
       })}
+      confirmActionsTracking={['add_endpoints', 'confirm']}
       hasFormError={isError}
       onSubmit={() => createEndpoint()}
       isSubmitPending={isPending}
@@ -213,6 +217,4 @@ const EndpointCreationPage: React.FC = () => {
       </FormField>
     </CreatePageLayout>
   );
-};
-
-export default EndpointCreationPage;
+}