+++ packages/manager/apps/vrack-services/src/pages/endpoints/endpoints.hook.ts
@@ -1,0 +1,22 @@
+import { useNavigate, useParams } from 'react-router-dom';
+import {
+  ButtonType,
+  PageLocation,
+  useOvhTracking,
+} from '@ovh-ux/manager-react-shell-client';
+import { urls } from '@/router/constants';
+
+export const useNavigateToCreateEndpointPage = () => {
+  const { trackClick } = useOvhTracking();
+  const navigate = useNavigate();
+  const { id } = useParams();
+
+  return () => {
+    trackClick({
+      location: PageLocation.page,
+      buttonType: ButtonType.button,
+      actions: ['create-endpoint'],
+    });
+    navigate(urls.createEndpoint.replace(':id', id));
+  };
+};