+++ packages/manager/apps/vrack-services/src/pages/subnets/Create.tsx
@@ -16,8 +16,8 @@ import {
   OsdsInput,
 } from '@ovhcloud/ods-components/react';
 import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
-import { useOvhTracking } from '@ovh-ux/manager-react-shell-client';
 import { ApiError, ApiResponse } from '@ovh-ux/manager-core-api';
+import { useOvhTracking, PageType } from '@ovh-ux/manager-react-shell-client';
 import {
   VrackServices,
   getVrackServicesResourceListQueryKey,
@@ -38,14 +38,13 @@ import {
   getSubnetCreationMutationKey,
   defaultCidr,
   defaultServiceRange,
-} from './constants';
+} from './subnets.constants';
 import { useVrackService } from '@/utils/vs-utils';
 import { FormField } from '@/components/FormField';
-import { urls, pageTrackingLabels } from '@/router/constants';
+import { urls } from '@/router/constants';
+import { PageName } from '@/utils/tracking';
 
-const dataTrackingPath = pageTrackingLabels[urls.createSubnet];
-
-const SubnetCreationPage: React.FC = () => {
+export default function SubnetCreationPage() {
   const { t } = useTranslation('vrack-services/subnets');
   const { id } = useParams();
   const [displayName, setDisplayName] = React.useState<string | undefined>(
@@ -93,12 +92,18 @@ const SubnetCreationPage: React.FC = () => {
         await queryClient.invalidateQueries({
           queryKey: getVrackServicesResourceQueryKey(id),
         }),
-        trackPage({ path: dataTrackingPath, value: '-succeess' }),
+        trackPage({
+          pageName: PageName.pendingCreateSubnet,
+          pageType: PageType.bannerInfo,
+        }),
       ]);
       navigate(dashboardUrl);
     },
     onError: () => {
-      trackPage({ path: dataTrackingPath, value: '-error' });
+      trackPage({
+        pageName: PageName.errorCreateSubnet,
+        pageType: PageType.bannerError,
+      });
     },
   });
 
@@ -113,11 +118,10 @@ const SubnetCreationPage: React.FC = () => {
       overviewUrl={dashboardUrl}
       title={t('createPageTitle')}
       createButtonLabel={t('createSubnetButtonLabel')}
-      dataTrackingPath={dataTrackingPath}
-      createButtonDataTracking="::confirm"
       formErrorMessage={t('subnetCreationError', {
         error: error?.response.data.message,
       })}
+      confirmActionsTracking={['add_subnets', 'confirm']}
       hasFormError={isError}
       onSubmit={() => createSubnet()}
       isSubmitPending={isPending}
@@ -227,6 +231,4 @@ const SubnetCreationPage: React.FC = () => {
       )}
     </CreatePageLayout>
   );
-};
-
-export default SubnetCreationPage;
+}