+++ packages/manager/modules/netapp/src/dashboard/module.js
@@ -9,12 +9,18 @@ import '@ovh-ux/ng-at-internet';
 
 import '@ovh-ux/manager-advices';
 import '@ovh-ux/manager-billing-components';
+import { ApiV2ListHelper } from '@ovh-ux/manager-ng-apiv2-helper';
+
 import component from './component';
+import service from './service';
 import routing from './routing';
 
 import index from './index/module';
 import snapshotPolicies from './snapshot-policies';
 import volumes from './volumes';
+import networkConfiguration from './network-configuration';
+import vrackConfiguration from './vrack-configuration';
+import deleteNetwork from './delete-network';
 
 const moduleName = 'ovhManagerNetAppDashboard';
 
@@ -29,12 +35,17 @@ angular
     'pascalprecht.translate',
     'ui.router',
     'ngOvhUtils',
+    ApiV2ListHelper.moduleName,
     snapshotPolicies,
     volumes,
+    networkConfiguration,
+    vrackConfiguration,
+    deleteNetwork,
     index,
   ])
   .config(routing)
   .component('ovhManagerNetAppDashboard', component)
+  .service('NetAppDashboardService', service)
   .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;