+++ packages/manager/apps/catalog/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-catalog-app@1.9.3...@ovh-ux/manager-catalog-app@1.9.4) (2024-04-17)
+
+**Note:** Version bump only for package @ovh-ux/manager-catalog-app
+
+
+
+
+
 ## [1.9.3](https://github.com/ovh/manager/compare/@ovh-ux/manager-catalog-app@1.9.2...@ovh-ux/manager-catalog-app@1.9.3) (2024-04-15)